refactor: ts, separate component, moved styles to own scss file, removed unused code, default to named exports
