+++ client/my-sites/email/email-management/style.scss
@@ -400,60 +400,6 @@
 	}
 }
 
-.email-plan-subscription__card.card.is-compact {
-	align-items: center;
-	display: flex;
-	flex-wrap: wrap;
-
-	@include break-small {
-		flex-wrap: initial;
-	}
-
-	.email-plan-subscription__renew {
-		margin-left: auto;
-		margin-right: 16px;
-
-		@include break-small {
-			flex: 1 1 auto;
-			margin-left: 16px;
-		}
-	}
-
-	.email-plan-subscription__auto-renew {
-		/* Ensure auto-renew text is using the standard font size */
-		label {
-			font-size: $font-body;
-		}
-
-		/* vertically aligns the auto-renew toggle - not sure if there's a better way... */
-		> div > div {
-			margin-top: auto;
-			margin-bottom: auto;
-		}
-	}
-}
-
-.email-plan-subscription__placeholder {
-	div {
-		@include placeholder( --color-neutral-5 );
-		margin: 0 auto;
-		width: 150px;
-		height: 28px;
-	}
-}
-
-.email-plan-subscription__description {
-	font-size: $font-body-small;
-
-	&.email-plan-subscription__description--expired {
-		color: var(--color-error-50);
-	}
-
-	.email-plan-subscription__description-renews-date {
-		margin-left: 20px;
-	}
-}
-
 .email-plan-warnings__container {
 	flex-basis: 100%;
 	margin: 5px 0;
